Add global screen size dialog and remove old dialog components
