--- apps/ccf-rui/src/app/modules/screen-size-notice/screen-size-notice.component.html
@@ -1,16 +1,0 @@
-<ng-container hraFeature="screen-size-notice">
-  <h3 mat-dialog-title class="header">
-    Heads up!
-    <button hraFeature="close" hraClickEvent mat-icon-button mat-dialog-close class="header-close">
-      <mat-icon>close</mat-icon>
-    </button>
-  </h3>
-
-  <mat-dialog-content class="content">
-    This website is optimized for Chrome or Firefox on a minimum resolution of 1280x832.
-  </mat-dialog-content>
-
-  <mat-dialog-actions class="buttons">
-    <button hraFeature="dismiss" hraClickEvent mat-button mat-dialog-close>Dismiss</button>
-  </mat-dialog-actions>
-</ng-container>
