--- apps/ccf-rui/src/app/modules/screen-size-notice/screen-size-notice.component.scss
@@ -1,48 +1,0 @@
-@use '@angular/material' as mat;
-@use 'vars';
-
-:host {
-  display: flex;
-  flex-direction: column;
-
-  .header {
-    font: vars.$headline-small;
-    letter-spacing: vars.$headline-small-tracking;
-    display: flex;
-    align-items: flex-end;
-    justify-content: space-between;
-
-    @include mat.dialog-overrides(
-      (
-        headline-padding: 0 0 0 2rem,
-      )
-    );
-
-    &::before {
-      display: none;
-    }
-
-    .header-close {
-      margin: 0.75rem;
-    }
-  }
-
-  mat-dialog-content {
-    font: vars.$body-large;
-    letter-spacing: vars.$body-large-tracking;
-    @include mat.dialog-overrides(
-      (
-        actions-padding: 1rem 2rem,
-        container-shape: 1rem,
-        supporting-text-color: #{vars.$primary},
-        with-actions-content-padding: 0 2rem,
-      )
-    );
-  }
-}
-
-::ng-deep {
-  .mat-mdc-dialog-container .mdc-dialog__surface {
-    box-shadow: unset;
-  }
-}
