+++ apps/cde-ui/src/app/app.component.ts
@@ -9,12 +9,6 @@ import { HraCommonModule, routeData } from '@hra-ui/common';
 import { ButtonsModule } from '@hra-ui/design-system/buttons';
 import { NavigationModule } from '@hra-ui/design-system/navigation';
 import { PlainTooltipDirective } from '@hra-ui/design-system/tooltips/plain-tooltip';
-import { ScreenSizeNoticeComponent } from './components/screen-size-notice/screen-size-notice.component';
-
-/** Max width to show screen size notice */
-export const SCREEN_SIZE_NOTICE_MAX_WIDTH = 1280;
-/** Max height to show screen size notice */
-export const SCREEN_SIZE_NOTICE_MAX_HEIGHT = 832;
 
 /**
  * App component for CDE
@@ -55,14 +49,6 @@ export class AppComponent extends BaseApplicationComponent {
 
   /** Header visibility (whether to show header or not) */
   protected readonly header = computed(() => this.data()['header'] as boolean | undefined);
-
-  /**
-   * Screen size notice detector of app component
-   */
-  protected readonly screenSizeNoticeDetector = ScreenSizeNoticeComponent.createDetector({
-    width: SCREEN_SIZE_NOTICE_MAX_WIDTH,
-    height: SCREEN_SIZE_NOTICE_MAX_HEIGHT,
-  });
 
   /** Initialize app */
   constructor() {
