--- apps/cde-ui/src/app/components/screen-size-notice/screen-size-notice.component.html
@@ -1,17 +1,0 @@
-<h3 mat-dialog-title class="header">
-  Heads up!
-  <button mat-icon-button mat-dialog-close class="header-close">
-    <mat-icon>close</mat-icon>
-  </button>
-</h3>
-
-<mat-dialog-content class="content">
-  This website is optimized for Chrome or Firefox on a minimum resolution of 1280x832.
-</mat-dialog-content>
-
-<mat-dialog-actions class="buttons">
-  <button mat-flat-button mat-dialog-close color="primary" class="atlas-btn cta-filled">
-    Continue
-    <mat-icon iconPositionEnd>east</mat-icon>
-  </button>
-</mat-dialog-actions>
