--- apps/cde-ui/src/app/components/screen-size-notice/screen-size-notice.component.scss
@@ -1,94 +1,0 @@
-@use '@angular/material' as mat;
-@use 'vars';
-
-:host {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  padding: 1.75rem 2rem;
-  max-width: 32.5rem;
-  box-shadow: 0px 5px 16px 0px #201e3d3d;
-
-  .header {
-    display: flex;
-    align-items: center;
-    height: 2rem;
-    margin-bottom: 0;
-
-    @include mat.dialog-overrides(
-      (
-        headline-padding: 0,
-        subhead-font: 'Metropolis',
-        subhead-size: 1.5rem,
-        subhead-weight: 500,
-      )
-    );
-
-    .header-close {
-      margin-left: auto;
-    }
-  }
-
-  mat-dialog-content {
-    width: 100%;
-    margin-bottom: 1rem;
-
-    @include mat.dialog-overrides(
-      (
-        container-shape: 1rem,
-        supporting-text-font: 'Nunito Sans Variable',
-        supporting-text-size: 1rem,
-        supporting-text-line-height: 1.5rem,
-        supporting-text-tracking: 0.5px,
-        supporting-text-color: vars.$on-secondary-fixed,
-        with-actions-content-padding: 0,
-      )
-    );
-  }
-
-  mat-dialog-actions {
-    @include mat.button-overrides(
-      (
-        filled-label-text-size: 0.875rem,
-        text-label-text-size: 0.875rem,
-      )
-    );
-
-    @include mat.dialog-overrides(
-      (
-        actions-alignment: space-between,
-        actions-padding: 0,
-      )
-    );
-  }
-
-  .atlas-btn {
-    padding: 0.5rem 1.5rem;
-    font-size: 1rem;
-    line-height: 1.5rem;
-    font-family: 'Metropolis';
-    height: 2.5rem;
-    letter-spacing: 0.5px;
-
-    .mat-icon {
-      margin-left: 0.75rem;
-      height: 1rem;
-      width: 1rem;
-      font-size: 1rem;
-    }
-  }
-
-  .btn-stroked {
-    @include mat.button-overrides(
-      (
-        text-horizontal-padding: 1rem,
-      )
-    );
-  }
-}
-
-::ng-deep {
-  .mat-mdc-dialog-container .mdc-dialog__surface {
-    box-shadow: unset;
-  }
-}
