+++ apps/ftu-ui/src/app/app.component.ts
@@ -1,11 +1,9 @@
 /* eslint-disable @angular-eslint/no-output-rename -- Allow rename for custom element events */
 import {
-  AfterContentInit,
   ChangeDetectionStrategy,
   Component,
   computed,
   HostBinding,
-  HostListener,
   inject,
   Input,
   model,
@@ -14,7 +12,6 @@ import {
   Output,
   SimpleChanges,
 } from '@angular/core';
-import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { MatDividerModule } from '@angular/material/divider';
 import { MatMenuModule } from '@angular/material/menu';
 import { ActivatedRoute, Router, RouterModule } from '@angular/router';
@@ -30,8 +27,6 @@ import {
   LinkRegistrySelectors,
   LinkType,
   ResourceRegistryActions,
-  StorageId,
-  StorageSelectors,
 } from '@hra-ui/cdk/state';
 import { HraCommonModule, routeData } from '@hra-ui/common';
 import { ButtonsModule } from '@hra-ui/design-system/buttons';
@@ -39,11 +34,7 @@ import { BreadcrumbItem } from '@hra-ui/design-system/buttons/breadcrumbs';
 import { IconsModule } from '@hra-ui/design-system/icons';
 import { NavigationModule } from '@hra-ui/design-system/navigation';
 import { PlainTooltipDirective } from '@hra-ui/design-system/tooltips/plain-tooltip';
-import {
-  FtuFullScreenService,
-  ScreenNoticeBehaviorComponent,
-  TissueLibraryBehaviorComponent,
-} from '@hra-ui/ftu-ui-components/src/lib/behavioral';
+import { FtuFullScreenService, TissueLibraryBehaviorComponent } from '@hra-ui/ftu-ui-components/src/lib/behavioral';
 import {
   FTU_DATA_IMPL_ENDPOINTS,
   FtuDataImplEndpoints,
@@ -73,6 +64,7 @@ import {
 } from '@hra-ui/state';
 import { Actions, ofActionDispatched } from '@ngxs/store';
 import { filter, from, map, Observable, OperatorFunction, ReplaySubject, switchMap, take } from 'rxjs';
+
 import { environment } from '../environments/environment';
 
 /** Input property keys */
@@ -90,9 +82,6 @@ type UpdateSelectors = Record<InputProps, boolean>;
 
 /** Link to main ftu page */
 export const ftuPage = createLinkId('FTU');
-
-/** Small view port size in pixels */
-const SMALL_VIEWPORT_THRESHOLD = 480; // In pixels
 
 /** Update selection where every part of the app should update  */
 const UPDATE_ALL_SELECTORS: UpdateSelectors = {
@@ -140,7 +129,7 @@ function filterUndefined<T>(): OperatorFunction<T | undefined, T> {
   },
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class AppComponent extends BaseApplicationComponent implements AfterContentInit, OnChanges, OnInit {
+export class AppComponent extends BaseApplicationComponent implements OnChanges, OnInit {
   /** Host binding of app component */
   @HostBinding('class.mat-typography') readonly matTypography = true;
 
@@ -393,54 +382,5 @@ export class AppComponent extends BaseApplicationComponent implements AfterConte
           this.updateSelectedIllustration();
         }
       });
-  }
-
-  /** Screen size notice open of app component */
-  screenSizeNoticeOpen = false;
-
-  /** Determines whether shown small viewport notice has been displayed */
-  private readonly hasShownSmallViewportNotice = selectQuerySnapshot(
-    StorageSelectors.get,
-    StorageId.Local,
-    'screen-size-notice',
-  );
-
-  /** Dialog  of app component */
-  private readonly dialog = inject(MatDialog);
-
-  /** Host listener for window resize events */
-  @HostListener('window:resize', ['$event'])
-  onWindowResize(): void {
-    this.detectSmallViewport();
-  }
-
-  /** Lifecycle hook that is called after content has been projected into the component */
-  ngAfterContentInit(): void {
-    this.detectSmallViewport();
-  }
-
-  /** Detect small viewport */
-  detectSmallViewport(): void {
-    if (
-      window.innerWidth <= SMALL_VIEWPORT_THRESHOLD &&
-      !this.hasShownSmallViewportNotice() &&
-      !this.screenSizeNoticeOpen
-    ) {
-      const dialogConfig: MatDialogConfig = {
-        disableClose: false,
-        panelClass: 'custom-overlay',
-        hasBackdrop: false,
-        minWidth: '19.5rem',
-      };
-
-      const ref = this.dialog.open(ScreenNoticeBehaviorComponent, dialogConfig);
-      ref.afterClosed().subscribe(() => (this.screenSizeNoticeOpen = false));
-      this.screenSizeNoticeOpen = true;
-    }
-
-    if (window.innerWidth > SMALL_VIEWPORT_THRESHOLD) {
-      this.screenSizeNoticeOpen = false;
-      this.dialog.closeAll();
-    }
   }
 }
