+++ apps/ftu-ui/src/assets/resources.yml
@@ -126,13 +126,6 @@ SizeLegend:
     - label: '100%'
       radius: 1.25
 
-ScreenSizeNoticeContent:
-  type: markdown
-  markdown: |
-    Functional Tissue Unit (FTU) Explorer is best utilized from a tablet or desktop.
-    The interface is not intended for this screen size. Proceed with this in mind, or,
-    read more about the FTU Explorer effort at the Human Reference Atlas (HRA) Portal.
-
 ContactAcknowledgement:
   type: text
   text: We received your message. Please allow two business days for a <br> response to each inquiry. Thanks for your time and expertise!
