+++ libs/application/package.json
@@ -5,7 +5,8 @@
     "@angular/core": "^20.2.0",
     "@hra-ui/common": "1.1.0",
     "zod": "^4.1.3",
-    "@hra-ui/design-system": "1.1.0"
+    "@hra-ui/design-system": "1.1.0",
+    "rxjs": "^7.8.2"
   },
   "sideEffects": false
 }
