+++ libs/ftu-ui-components/src/lib/behavioral/src/index.ts
@@ -1,6 +1,5 @@
 export * from './lib/biomarker-details/biomarker-details.component';
 export * from './lib/medical-illustration-behavior/medical-illustration-behavior.component';
 export * from './lib/tissue-library-behavior/tissue-library-behavior.component';
-export * from './lib/screen-notice-behavior/screen-notice-behavior.component';
 export * from './lib/biomarker-details-wc/biomarker-details-wc.component';
 export * from './lib/ftu-fullscreen-service/ftu-fullscreen.service';
