--- libs/ftu-ui-components/src/lib/behavioral/src/lib/screen-notice-behavior/screen-notice-behavior.component.html
@@ -1,1 +1,0 @@
-<!-- This component opens a dialog programmatically using DialogService -->
