--- libs/ftu-ui-components/src/lib/behavioral/src/lib/screen-notice-behavior/screen-notice-behavior.component.scss
@@ -1,3 +1,0 @@
-:host {
-  display: block;
-}
