+++ libs/ftu-ui-components/src/lib/molecules/src/index.ts
@@ -1,6 +1,5 @@
 export * from './lib/biomarker-table-data-icon/biomarker-table-data-icon.component';
 export * from './lib/fullscreen/fullscreen-container.component';
 export * from './lib/interactive-svg/interactive-svg.component';
-export * from './lib/screen-size-notice/screen-size-notice.component';
 export * from './lib/source-list/source-list.component';
 export * from './lib/tissue-tree-list/tissue-tree-list.component';
