--- libs/ftu-ui-components/src/lib/molecules/src/lib/screen-size-notice/screen-size-notice.component.html
@@ -1,9 +1,0 @@
-<h3>Screen Size Notice</h3>
-<button mat-icon-button class="header-close" (click)="proceedClick.emit()">
-  <mat-icon class="material-symbols-outlined">close</mat-icon>
-</button>
-<markdown [data]="content" class="content"></markdown>
-<button mat-stroked-button color="accent" class="btn-stroked" (click)="proceedClick.emit()">
-  Proceed to FTU Explorer
-</button>
-<button mat-flat-button color="primary" class="btn-flat" [hraLink]="portal">Human Reference Atlas Portal</button>
