--- libs/ftu-ui-components/src/lib/molecules/src/lib/screen-size-notice/screen-size-notice.component.scss
@@ -1,40 +1,0 @@
-:host {
-  display: none;
-  position: relative;
-  border-radius: 0.5rem;
-  padding: 1.5rem;
-  flex-direction: column;
-  max-width: 30rem;
-
-  .header-close {
-    position: absolute;
-    top: 1.125rem;
-    right: 1rem;
-  }
-
-  h3 {
-    margin: 0;
-    height: 2.5rem;
-    line-height: 2.5rem;
-  }
-
-  button {
-    height: 3rem;
-    padding: 0.75rem 0;
-    letter-spacing: 0.5px;
-  }
-
-  .btn-stroked {
-    margin-bottom: 1rem;
-  }
-
-  .content {
-    font-size: 1rem;
-    line-height: 1.5rem;
-    letter-spacing: 0.5px;
-  }
-
-  @media (max-width: 1279px) {
-    display: flex;
-  }
-}
